+++ openbank-engagement-service/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.12.1](https://github.com/JiRaska/open-bank-oss/compare/engagement-service-v0.12.0...engagement-service-v0.12.1) (2026-08-18)
+
+
+### Bug Fixes
+
+* **libs-runtime:** switch AbstractOutboxDispatcher.metrics to constructor injection ([#5199](https://github.com/JiRaska/open-bank-oss/issues/5199)) ([1d07563](https://github.com/JiRaska/open-bank-oss/commit/1d075635ef70004e8b9b50475cd97aa31c9beafd)), closes [#5128](https://github.com/JiRaska/open-bank-oss/issues/5128)
+
 ## [0.12.0](https://github.com/JiRaska/open-bank-oss/compare/engagement-service-v0.11.1...engagement-service-v0.12.0) (2026-08-17)